refactor: remove deployment target card base component
