+++ frontend/ui/src/app/deployments/deployment-target-card/deployment-target-dashboard-card.component.html
@@ -80,11 +80,3 @@
     </div>
   }
 </div>
-
-<!-- TODO: Refactor these viewChild properties and related features out of the base component or make them non-required -->
-<ng-template #instructionsModal />
-<ng-template #deploymentModal />
-<ng-template #deleteConfirmModal />
-<ng-template #manageDeploymentTargetDrawer />
-<ng-template #deploymentTargetNotesDrawer />
-<ng-template #deleteDeploymentProgressModal />
