+++ frontend/ui/src/app/deployments/deployment-target-card/deployment-target-dashboard-card.component.ts
@@ -1,15 +1,17 @@
 import {OverlayModule} from '@angular/cdk/overlay';
 import {TextFieldModule} from '@angular/cdk/text-field';
 import {AsyncPipe, NgOptimizedImage} from '@angular/common';
-import {Component} from '@angular/core';
+import {Component, input} from '@angular/core';
 import {ReactiveFormsModule} from '@angular/forms';
 import {RouterLink} from '@angular/router';
+import {DeploymentTarget} from '@distr-sh/distr-sdk';
 import {FaIconComponent} from '@fortawesome/angular-fontawesome';
+import {faArrowUpRightFromSquare} from '@fortawesome/free-solid-svg-icons';
 import {SecureImagePipe} from '../../../util/secureImage';
 import {DeploymentTargetStatusDotComponent} from '../../components/status-dot';
+import {DeploymentTargetLatestMetrics} from '../../types/deployment-target-metrics';
 import {DeploymentAppNameComponent} from './deployment-app-name.component';
 import {DeploymentStatusTextComponent} from './deployment-status-text.component';
-import {DeploymentTargetCardBaseComponent} from './deployment-target-card-base.component';
 import {DeploymentTargetMetricsComponent} from './deployment-target-metrics.component';
 
 @Component({
@@ -30,4 +32,9 @@ import {DeploymentTargetMetricsComponent} from './deployment-target-metrics.comp
     RouterLink,
   ],
 })
-export class DeploymentTargetDashboardCardComponent extends DeploymentTargetCardBaseComponent {}
+export class DeploymentTargetDashboardCardComponent {
+  public readonly deploymentTarget = input.required<DeploymentTarget>();
+  public readonly deploymentTargetMetrics = input<DeploymentTargetLatestMetrics | undefined>(undefined);
+
+  protected readonly faArrowUpRightFromSquare = faArrowUpRightFromSquare;
+}
